requirements.txt
hyperopt>=0.1.2
numba>=0.46.0
numpy>=1.14.5
tensorflow-gpu>=1.10.0
tifffile==2020.2.16
tqdm>=4.36.1
scikit-learn>=0.22.1
scipy~=1.4.1
matplotlib~=3.3.0